Security Considerations section
Security Considerations, one typo.

Adding a section on Security Considerations.
Since we are still in discussions around security considerations for the API, I left this one somewhat ambiguous.  

This is by no means perfect, but now we have something.  Trying to get the draft into a minimal reasonable state before we submit for March.

from discussions in:
https://github.com/bgp/draft-ietf-peering-api/issues/6
https://github.com/bgp/draft-ietf-peering-api/issues/4