Rework security considerations section
This commit delves more into the security considerations to operate the
peering API and introduces a baseline threat model that supports the
addition of authorization, proof of resource holdership and request
integrity to address them.

This commit also moves the security considerations sections to the end of the
document to align with other RFCs.